Made requested changes to TimeService and ResourceService

+++ DotNetTribes/DTOs/ResourceDto.cs
@@ -7,6 +7,6 @@ namespace DotNetTribes.DTOs
     {
         public string Type { get; set; }
         public int Amount { get; set; }
-        public int UpdatedAt { get; set; }
+        public long UpdatedAt { get; set; }
     }
 }
